+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Can see the unlocking amount transaction when an order is matched on the price less than was locked
 * 500 on get history (create manage limits request)
+* participants statistics for immediate sale
 
 ## 3.7.1